# docs/swagger.yaml
basePath: /api
definitions:
...
  model.Menu:
    properties:
      children:
        items:
          $ref: '#/definitions/model.Menu'
        type: array
      created_at:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      path:
        type: string
      permission:
        type: string
      sort:
        type: integer
      status:
        type: integer
      type:
        type: integer
      updated_at:
        type: string
    type: object
  model.Role:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      menus:
        items:
          $ref: '#/definitions/model.Menu'
        type: array
      name:
        type: string
      status:
        type: integer
      updated_at:
        type: string
    type: object
  model.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      phone:
        type: string
      roles:
        items:
          $ref: '#/definitions/model.Role'
        type: array
      status:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  utils.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: 后台管理系统 API 文档
  title: GinAdmin API
  version: "1.0"
paths:
...
  /menus:
    get:
      description: 获取完整菜单树形结构
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Menu'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 获取菜单树
      tags:
      - 菜单管理
    post:
      consumes:
      - application/json
      description: 创建新菜单（目录/菜单页/按钮权限）
      parameters:
      - description: 菜单信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 新建菜单
      tags:
      - 菜单管理
  /menus/{id}:
    delete:
      description: 根据 ID 删除菜单，有子菜单时拒绝删除
      parameters:
      - description: 菜单 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 存在子菜单，无法删除
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 删除菜单
      tags:
      - 菜单管理
    get:
      description: 根据 ID 查询菜单详情
      parameters:
      - description: 菜单 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 菜单不存在
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 获取菜单详情
      tags:
      - 菜单管理
    put:
      consumes:
      - application/json
      description: 更新菜单信息
      parameters:
      - description: 菜单 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 菜单信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 编辑菜单
      tags:
      - 菜单管理
  /roles:
    get:
      description: 分页查询角色列表，支持关键词搜索
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 搜索关键词（角色名/编码）
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  properties:
                    list:
                      items:
                        $ref: '#/definitions/model.Role'
                      type: array
                    total:
                      type: integer
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: 获取角色分页列表
      tags:
      - 角色管理
    post:
      consumes:
      - application/json
      description: 创建新角色并分配菜单权限
      parameters:
      - description: 角色信息
        in: body
        name: body
        required: true
        schema:
          type: object
      - description: 角色名称
        in: body
        name: body.body.name
        required: true
        schema:
          type: string
      - description: 角色编码
        in: body
        name: body.body.code
        required: true
        schema:
          type: string
      - description: 描述
        in: body
        name: body.body.description
        schema:
          type: string
      - description: 菜单 ID 列表
        in: body
        name: body.body.menu_ids
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 新建角色
      tags:
      - 角色管理
  /roles/{id}:
    delete:
      description: 根据 ID 删除角色
      parameters:
      - description: 角色 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 删除角色
      tags:
      - 角色管理
    get:
      description: 根据 ID 查询角色详情，包含关联的菜单 ID 列表
      parameters:
      - description: 角色 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 角色不存在
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 获取角色详情
      tags:
      - 角色管理
    put:
      consumes:
      - application/json
      description: 更新角色信息及菜单权限
      parameters:
      - description: 角色 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 角色信息
        in: body
        name: body
        required: true
        schema:
          type: object
      - description: 角色名称
        in: body
        name: body.body.name
        required: true
        schema:
          type: string
      - description: 角色编码
        in: body
        name: body.body.code
        required: true
        schema:
          type: string
      - description: 描述
        in: body
        name: body.body.description
        schema:
          type: string
      - description: 菜单 ID 列表
        in: body
        name: body.body.menu_ids
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 编辑角色
      tags:
      - 角色管理
  /users:
    get:
      description: 分页查询用户列表，支持关键词搜索
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 搜索关键词（用户名/邮箱）
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  properties:
                    list:
                      items:
                        $ref: '#/definitions/model.User'
                      type: array
                    total:
                      type: integer
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: 获取用户分页列表
      tags:
      - 用户管理
    post:
      consumes:
      - application/json
      description: 创建新用户并分配角色
      parameters:
      - description: 用户信息
        in: body
        name: body
        required: true
        schema:
          type: object
      - description: 用户名
        in: body
        name: body.body.username
        required: true
        schema:
          type: string
      - description: 密码
        in: body
        name: body.body.password
        required: true
        schema:
          type: string
      - description: 邮箱
        in: body
        name: body.body.email
        schema:
          type: string
      - description: 手机号
        in: body
        name: body.body.phone
        schema:
          type: string
      - default: 1
        description: 状态 1=启用 0=禁用
        in: body
        name: body.body.status
        schema:
          type: integer
      - description: 角色 ID 列表
        in: body
        name: body.body.role_ids
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 新建用户
      tags:
      - 用户管理
  /users/{id}:
    delete:
      description: 根据 ID 删除用户
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 删除用户
      tags:
      - 用户管理
    get:
      description: 根据 ID 查询用户详情
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 用户不存在
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 获取用户详情
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      description: 更新用户信息，密码为空时不覆盖
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户信息
        in: body
        name: body
        required: true
        schema:
          type: object
      - description: 用户名
        in: body
        name: body.body.username
        required: true
        schema:
          type: string
      - description: 密码（为空不修改）
        in: body
        name: body.body.password
        schema:
          type: string
      - description: 邮箱
        in: body
        name: body.body.email
        schema:
          type: string
      - description: 手机号
        in: body
        name: body.body.phone
        schema:
          type: string
      - description: 状态 1=启用 0=禁用
        in: body
        name: body.body.status
        schema:
          type: integer
      - description: 角色 ID 列表
        in: body
        name: body.body.role_ids
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 编辑用户
      tags:
      - 用户管理
securityDefinitions:
  BearerAuth:
    description: 'JWT Token，格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"